created target estimator and made turning work
