+++ components/target_estimator.py
@@ -1,89 +1,21 @@
-import math
-
-from networktables import NetworkTables
-from typing import Optional
-from dataclasses import dataclass
-from magicbot import feedback
-from wpilib import Timer
+from vision import Vision
+from typing import Optional, Tuple
 
 
-@dataclass
-class VisionData:
-    #: The hy to the target in metres.
-    x: float
+class TargetEsitmator:
+    """Fuses data from vision, odometry, the imu and other sources
+    to estimate where the target is from the robot and where you should aim to hit it when moving"""
 
-    #: The angle to the target in radians.
-    y: float
-
-    #: An arbitrary timestamp, in seconds,
-    #: for when the vision system last obtained data.
-    timestamp: float
-
-    __slots__ = ("x", "y", "timestamp")
-
-
-class TargetEstimator:
-    """Combines data from vision, odometry, the imu and other sources
-    to estimate where the target is relative to the robot
-    """
-
-    SYSTEM_LAG_THRESHOLD = 0.200
+    # https://robotpy.readthedocs.io/projects/wpimath/en/stable/wpimath.estimator.html
+    vision: Vision
 
     def __init__(self) -> None:
-
-        self.nt = NetworkTables
-        self.table = self.nt.getTable("/vision")
-        self.vision_data_entry = self.table.getEntry("data")
-        self.ping_time_entry = self.table.getEntry("ping")
-        self.rio_pong_time_entry = self.table.getEntry("rio_pong")
-        self.raspi_pong_time_entry = self.table.getEntry("raspi_pong")
-        self.latency_entry = self.table.getEntry("clock_offset")
-
-        self.last_pong = Timer.getFPGATimestamp()
-        self.latency = 0
-
-        self.vision_data = None
-
-    def get_data(self) -> Optional[VisionData]:
-        """Returns the latest vision data.
-
-        Returns None if there is no vision data.
-        """
-        return self.vision_data
+        self.angle_to_target: Optional[float] = None
+        self.position: Optional[Tuple[float, float]] = None
 
     def execute(self) -> None:
+        visionData = self.vision.get_data()
+        self.angle_to_target = visionData.angle
 
-        data = self.vision_data_entry.getDoubleArray(None)
-        if data is not None:
-            self.vision_data = VisionData(
-                data[0], data[1], data[2] + self.get_latency()
-            )
-            # add latency to vision timestamp
-
-        self.nt.flush()
-
-    @feedback
-    def is_ready(self) -> bool:
-        return self.system_lag_calculation() < self.SYSTEM_LAG_THRESHOLD
-
-    @feedback
-    def system_lag_calculation(self) -> float:
-        if self.vision_data is not None:
-            return Timer.getFPGATimestamp() - self.vision_data.timestamp
-        else:
-            return math.inf
-
-    def ping(self) -> None:
-        """Send a ping to the RasPi to determine the connection latency."""
-        self.ping_time_entry.setDouble(Timer.getFPGATimestamp())
-
-    def pong(self) -> None:
-        """Receive a pong from the RasPi to determine the connection latency."""
-        rio_pong_time = self.rio_pong_time_entry.getDouble(0)
-        if abs(rio_pong_time - self.last_pong) > 1e-4:  # Floating point comparison
-            raspi_pong_time = self.raspi_pong_time_entry.getDouble(0)
-            self.latency_entry.setDouble(rio_pong_time - raspi_pong_time)
-            self.last_pong = rio_pong_time
-
-    def get_latency(self):
-        return self.latency_entry.getDouble(0)
+    def to_target(self) -> float:
+        return self.angle_to_target
