+++ robot.py
@@ -12,6 +12,7 @@ from components.indexer import Indexer
 from components.intake import Intake
 from components.shooter import Shooter
 from components.turret import Turret
+from components.vision import Vision
 from components.target_estimator import TargetEstimator
 
 from controllers.shooter import ShooterController
@@ -31,6 +32,7 @@ class MyRobot(magicbot.MagicRobot):
     indexer: Indexer
     shooter: Shooter
     turret: Turret
+    vision: Vision
     target_estimator: TargetEstimator
 
     def createObjects(self):
